Realtime coloring enhancement state/color change in other view (#1073)
